requirements.txt
pytest
moviepy
pydub
pillow
scipy
matplotlib
opencv-python
dataclasses-json
tqdm
flask
requests
websocket-client
